+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 6.5.1 (in progress)
 
 ##### New Features
-* [#2603](https://github.com/oshi/oshi/pull/2603): Add part number to Global Memory - [@BartekDziurowicz](https://github.com/BartekDziurowicz).
+* [#2603](https://github.com/oshi/oshi/pull/2603),
+  [#2603](https://github.com/oshi/oshi/pull/2625): Add part number to Physical Memory - [@BartekDziurowicz](https://github.com/BartekDziurowicz), [@dbwiddis](https://github.com/dbwiddis).
 
 ##### Bug fixes / Improvements
 * [#2605](https://github.com/oshi/oshi/pull/2605): Reduce CpuStat.getSystemCpuLoadticks memory allocation pressure - [@chrisribble](https://github.com/chrisribble).